Fix issue where AC states would flap
Fix an issue where the AC states would revert back to a previous
state in HomeAssistant. Issue was caused by the 'control' update
occurring during a poll state, this results in the stale polled
state to be published instead of the true latest state.

Solution is to force a poll restart after each 'control' update.
This ensures that the next publish will contain the most recent
unit states.