Remove AUTO mode
The MSXY-FN10VE does not support AUTO mode.
If needed, explicitly expose the MODE as HEAT_COOL instead of AUTO.

If the A/C is set to another mode such as FAN_ONLY or DRY,
changing it to AUTO mode causes the A/C to return an unknown
MODE value of 0x00.